New feature: filter jobs automatically when the user clicks the button
You no longer have to edit regular expressions in the extension
options to hide jobs. Now when you click the button to tell LinkedIn
not to show you a job anymore (the eye with the slash through it), the
extension will remember that automatically, and it'll subsequently
hide the job reliably even if LinkedIn doesn't.