Handle changes to LinkedIn web app page format
LinkedIn has changed the format of the job listings on some of their
pages, so we need to update our code to accommodate the new format.

Most notably, the workplace type (On-site, Remote, Hybrid) is now
sometimes a separate HTML element, and the descriptive label for the
dismiss job button is different sometimes.