Embed utils.js code in other files that need it, instead of importing
When an import was being used, then whenever the extension was updated
in a running Chrome instance, utils.js started failing to import with
the error "Failed to fetch dynamically imported module." I spent hours
banging my head against this trying to figure out what's causing it
and how to fix it and eventually gave up. The path of least resistence
is to just embed a copy of the code in all the files that need it.