Save base64 approach to button URL for historical purposes
I spent hours figuring out how to convert our 16-pixel icon into a
base64 URL so I could embed it into the LinkedIn page because
embedding the chrome-extension:// URL wasn't working, and then
realized that the only reason it wasn't working was because I hadn't
added it to web_accessibl_resources in the manifest, which is
obviously the correct solution. I'm committing this for historical
reference, because there are a couple of clever JS tricks in here that
I might need to refer back to later, but I'm not actually going to use
it.