Go back to using a base64 URL for the button image
This reverts commit c8e4d59457260c4330497d1fcac1b451f028c232.

I don't know why, but for some reason when the extension is reloaded,
the buttom image stops loading properly until the browser is
restarted. Perhaps this is only a problem when the extension is
unpacked :shrug:, but regardless, the easiest workaround is to use use
the baes64 URL. All my work to figure out how to generate that
apparently has paid off!