Add private hide functionality, plus a ton of other changes
This is a huge, monolithic commit because I've been working on it for
12+ hours virtually non-stop and all the changes are commingled and I
don't have the time or energy to tease them about and commit them
separately.

Functional changes:

"Private" job hiding is now supported. You can now tell tell the
extension to hide a particular job without telling LinkedIn that. This
is the right thing to do if a job is a good fit for you but you don't
want to see that specific job again, e.g., because you've already
applied to it so there's no need. This way you won't inadvertently
train LinkedIn's algorithm to show you fewer jobs like it in the
future.

The extension's functionality now remains "wired into" displayed jobs
no matter how many times you click the hide or undo hide buttons;
previously once you clicked them once the extension stopped doing
anything with that job until you refreshed the page. I don't know why
you'd want to keep clicking hide / unhide / hide / unhide multiple
times, but if you do for whatever reason, the extension now honors
those requests.

Previously there was a two-second delay between when you unhid a job
and when it was removed from the list of hidden jobs in the options.
Now this happens much more quickly, essentially as soon as the job
finishes being unhidden.

Backend changes:

To cope with the fact that the LinkedIn code randomly replaces job
cards, we now maintain a cache of job states indexed on the job title,
company and location, rather than only linking job states to the
cards. This allows us to keep the extension in sync with the LinkedIn
app, restoring listeners and buttons as needed when they disappear
because the LinkedIn app is moving things around.

A bunch of debug logging, disabled in production builds, has been
added, including the ability to add unique identifiers to arbitrary
objects so those identifiers can be included in debug logs to
facilitate debugging things like LinkedIn randomly replacing job
cards.

The logic for how and when to modify the DOM has been transformed into
a state machine with extensive documentation in the code.