Don't save options when nothing has changed
When we updated the code to split long lists in options into multiple
key/value pairs in storage to avoid value length limits, we missed
updating the code for saving options that checks and only saves the
things that have changed. This is now fixed. Also, I added a unit test
to ensure this doesn't break again.