Log classes of matched elements in classes.log while running tests
To assist us in eventually removing obsolete DOM element classes from
our code base, log matched classes while running tests, so that we
will over time collect a history of when classes are used. This way we
can look back and remove from the code bases any classes that have not
appeared in a sufficiently long time, indicating that they are
obsolete.